add sqlite

+++ SweetsDokkana/SweetsDokkana/Models/Product.cs
@@ -9,7 +9,7 @@ namespace SweetsDokkana.Models
         public int Id { get; set; }
         public string discreption { get; set; }
         public string name { get; set; }
-        public int price { get; set; }
+        public float price { get; set; }
         public string image_link { get; set; }
     }
 }
